Explicitly set background color in tmux
When starting tmux, query the terminal for its current background color
and explicitly set tmux's background to that. This is necessary so that
tmux will return the proper color when queried by neovim.

If a palette index is used, or if the bg is set to 'terminal', tmux will
return the default xterm color at that index rather than the color
that's actually being displayed.